feat(puppeteer): network traffics manipulation
