+++ lib/helper/Playwright.js
@@ -50,8 +50,9 @@ const { createValueEngine, createDisabledEngine } = require('./extras/Playwright
 const {
   seeElementError, dontSeeElementError, dontSeeElementInDOMError, seeElementInDOMError,
 } = require('./errors/ElementAssertion');
-const { createAdvancedTestResults, allParameterValuePairsMatchExtreme, extractQueryObjects } = require('./networkTraffics/utils');
-const { log } = require('../output');
+const {
+  dontSeeTraffic, seeTraffic, grabRecordedNetworkTraffics, stopRecordingTraffic, flushNetworkTraffics,
+} = require('./networkTraffics/shares/actions');
 
 const pathSeparator = path.sep;
 
@@ -3008,37 +3009,6 @@ class Playwright extends Helper {
   }
 
   /**
-   * {{> grabRecordedNetworkTraffics }}
-   */
-  async grabRecordedNetworkTraffics() {
-    if (!this.recording || !this.recordedAtLeastOnce) {
-      throw new Error('Failure in test automation. You use "I.grabRecordedNetworkTraffics", but "I.startRecordingTraffic" was never called before.');
-    }
-
-    const promises = this.requests.map(async (request) => {
-      const resp = await request.response;
-      let body;
-      try {
-        // There's no 'body' for some requests (redirect etc...)
-        body = JSON.parse((await resp.body()).toString());
-      } catch (e) {
-        // only interested in JSON, not HTML responses.
-      }
-
-      return {
-        url: resp.url(),
-        response: {
-          status: resp.status(),
-          statusText: resp.statusText(),
-          body,
-        },
-      };
-    });
-
-    return Promise.all(promises);
-  }
-
-  /**
    * Blocks traffic of a given URL or a list of URLs.
    *
    * Examples:
@@ -3117,170 +3087,46 @@ class Playwright extends Helper {
   }
 
   /**
+   *
    * {{> flushNetworkTraffics }}
    */
   flushNetworkTraffics() {
-    this.requests = [];
+    flushNetworkTraffics.call(this);
   }
 
   /**
+   *
    * {{> stopRecordingTraffic }}
    */
   stopRecordingTraffic() {
-    this.page.removeAllListeners('request');
-    this.recording = false;
+    stopRecordingTraffic.call(this);
   }
 
   /**
+   *
+   * {{> grabRecordedNetworkTraffics }}
+   */
+  async grabRecordedNetworkTraffics() {
+    return grabRecordedNetworkTraffics.call(this);
+  }
+
+  /**
+   *
    * {{> seeTraffic }}
    */
   async seeTraffic({
     name, url, parameters, requestPostData, timeout = 10,
   }) {
-    if (!name) {
-      throw new Error('Missing required key "name" in object given to "I.seeTraffic".');
-    }
-
-    if (!url) {
-      throw new Error('Missing required key "url" in object given to "I.seeTraffic".');
-    }
-
-    if (!this.recording || !this.recordedAtLeastOnce) {
-      throw new Error('Failure in test automation. You use "I.seeTraffic", but "I.startRecordingTraffic" was never called before.');
-    }
-
-    for (let i = 0; i <= timeout * 2; i++) {
-      const found = this._isInTraffic(url, parameters);
-      if (found) {
-        return true;
-      }
-      await new Promise((done) => {
-        setTimeout(done, 1000);
-      });
-    }
-
-    // check request post data
-    if (requestPostData && this._isInTraffic(url)) {
-      const advancedTestResults = createAdvancedTestResults(url, requestPostData, this.requests);
-
-      assert.equal(advancedTestResults, true, `Traffic named "${name}" found correct URL ${url}, BUT the post data did not match:\n ${advancedTestResults}`);
-    } else if (parameters && this._isInTraffic(url)) {
-      const advancedTestResults = createAdvancedTestResults(url, parameters, this.requests);
-
-      assert.fail(
-        `Traffic named "${name}" found correct URL ${url}, BUT the query parameters did not match:\n`
-        + `${advancedTestResults}`,
-      );
-    } else {
-      assert.fail(
-        `Traffic named "${name}" not found in recorded traffic within ${timeout} seconds.\n`
-        + `Expected url: ${url}.\n`
-        + `Recorded traffic:\n${this._getTrafficDump()}`,
-      );
-    }
+    await seeTraffic.call(this, ...arguments);
   }
 
   /**
-   * Returns full URL of request matching parameter "urlMatch".
    *
-   * @param {string|RegExp} urlMatch Expected URL of request in network traffic. Can be a string or a regular expression.
-   *
-   * Examples:
-   *
-   * ```js
-   * I.grabTrafficUrl('https://api.example.com/session');
-   * I.grabTrafficUrl(/session.*start/);
-   * ```
-   *
-   * @return {Promise<*>}
-   */
-  grabTrafficUrl(urlMatch) {
-    if (!this.recordedAtLeastOnce) {
-      throw new Error('Failure in test automation. You use "I.grabTrafficUrl", but "I.startRecordingTraffic" was never called before.');
-    }
-
-    for (const i in this.requests) {
-      // eslint-disable-next-line no-prototype-builtins
-      if (this.requests.hasOwnProperty(i)) {
-        const request = this.requests[i];
-
-        if (request.url && request.url.match(new RegExp(urlMatch))) {
-          return request.url;
-        }
-      }
-    }
-
-    assert.fail(`Method "getTrafficUrl" failed: No request found in traffic that matches ${urlMatch}`);
-  }
-
-  /**
    * {{> dontSeeTraffic }}
    *
    */
   dontSeeTraffic({ name, url }) {
-    if (!this.recordedAtLeastOnce) {
-      throw new Error('Failure in test automation. You use "I.dontSeeTraffic", but "I.startRecordingTraffic" was never called before.');
-    }
-
-    if (!name) {
-      throw new Error('Missing required key "name" in object given to "I.dontSeeTraffic".');
-    }
-
-    if (!url) {
-      throw new Error('Missing required key "url" in object given to "I.dontSeeTraffic".');
-    }
-
-    if (this._isInTraffic(url)) {
-      assert.fail(`Traffic with name "${name}" (URL: "${url}') found, but was not expected to be found.`);
-    }
-  }
-
-  /**
-   * Checks if URL with parameters is part of network traffic. Returns true or false. Internal method for this helper.
-   *
-   * @param url URL to look for.
-   * @param [parameters] Parameters that this URL needs to contain
-   * @return {boolean} Whether or not URL with parameters is part of network traffic.
-   * @private
-   */
-  _isInTraffic(url, parameters) {
-    let isInTraffic = false;
-    this.requests.forEach((request) => {
-      if (isInTraffic) {
-        return; // We already found traffic. Continue with next request
-      }
-
-      if (!request.url.match(new RegExp(url))) {
-        return; // url not found in this request. continue with next request
-      }
-
-      // URL has matched. Now we check the parameters
-
-      if (parameters) {
-        const advancedReport = allParameterValuePairsMatchExtreme(extractQueryObjects(request.url), parameters);
-        if (advancedReport === true) {
-          isInTraffic = true;
-        }
-      } else {
-        isInTraffic = true;
-      }
-    });
-
-    return isInTraffic;
-  }
-
-  /**
-   * Returns all URLs of all network requests recorded so far during execution of test scenario.
-   *
-   * @return {string} List of URLs recorded as a string, separated by new lines after each URL
-   * @private
-   */
-  _getTrafficDump() {
-    let dumpedTraffic = '';
-    this.requests.forEach((request) => {
-      dumpedTraffic += `${request.method} - ${request.url}\n`;
-    });
-    return dumpedTraffic;
+    dontSeeTraffic.call(this, ...arguments);
   }
 
   /**
