+++ lib/helper/WebDriver.js
@@ -2,10 +2,8 @@ let webdriverio;
 
 const assert = require('assert');
 const path = require('path');
-const fs = require('fs');
 
 const Helper = require('@codeceptjs/helper');
-const crypto = require('crypto');
 const promiseRetry = require('promise-retry');
 const stringIncludes = require('../assert/include').includes;
 const { urlEquals, equals } = require('../assert/equal');
@@ -30,13 +28,14 @@ const ElementNotFound = require('./errors/ElementNotFound');
 const ConnectionRefused = require('./errors/ConnectionRefused');
 const Locator = require('../locator');
 const { highlightElement } = require('./scripts/highlightElement');
-const store = require('../store');
 const { focusElement } = require('./scripts/focusElement');
 const { blurElement } = require('./scripts/blurElement');
 const {
   dontSeeElementError, seeElementError, seeElementInDOMError, dontSeeElementInDOMError,
 } = require('./errors/ElementAssertion');
-const { allParameterValuePairsMatchExtreme, extractQueryObjects, createAdvancedTestResults } = require('./networkTraffics/utils');
+const {
+  dontSeeTraffic, seeTraffic, grabRecordedNetworkTraffics, stopRecordingTraffic, flushNetworkTraffics,
+} = require('./networkTraffics/shares/actions');
 
 const SHADOW = 'shadow';
 const webRoot = 'body';
@@ -2758,42 +2757,7 @@ class WebDriver extends Helper {
       console.log('* Switch to devtools protocol to use this command by setting devtoolsProtocol: true in the configuration');
       return;
     }
-    if (!this.recording || !this.recordedAtLeastOnce) {
-      throw new Error('Failure in test automation. You use "I.grabRecordedNetworkTraffics", but "I.startRecordingTraffic" was never called before.');
-    }
-
-    const promises = this.requests.map(async (request) => {
-      const resp = await request.response;
-
-      if (!resp) {
-        return {
-          url: '',
-          response: {
-            status: '',
-            statusText: '',
-            body: '',
-          },
-        };
-      }
-
-      let body;
-      try {
-        // There's no 'body' for some requests (redirect etc...)
-        body = JSON.parse((await resp.body()).toString());
-      } catch (e) {
-        // only interested in JSON, not HTML responses.
-      }
-
-      return {
-        url: resp.url(),
-        response: {
-          status: resp.status(),
-          statusText: resp.statusText(),
-          body,
-        },
-      };
-    });
-    return Promise.all(promises);
+    return grabRecordedNetworkTraffics.call(this);
   }
 
   /**
@@ -2809,47 +2773,7 @@ class WebDriver extends Helper {
       console.log('* Switch to devtools protocol to use this command by setting devtoolsProtocol: true in the configuration');
       return;
     }
-    if (!name) {
-      throw new Error('Missing required key "name" in object given to "I.seeTraffic".');
-    }
-
-    if (!url) {
-      throw new Error('Missing required key "url" in object given to "I.seeTraffic".');
-    }
-
-    if (!this.recording || !this.recordedAtLeastOnce) {
-      throw new Error('Failure in test automation. You use "I.seeTraffic", but "I.startRecordingTraffic" was never called before.');
-    }
-
-    for (let i = 0; i <= timeout * 2; i++) {
-      const found = this._isInTraffic(url, parameters);
-      if (found) {
-        return true;
-      }
-      await new Promise((done) => {
-        setTimeout(done, 1000);
-      });
-    }
-
-    // check request post data
-    if (requestPostData && this._isInTraffic(url)) {
-      const advancedTestResults = createAdvancedTestResults(url, requestPostData, this.requests);
-
-      assert.equal(advancedTestResults, true, `Traffic named "${name}" found correct URL ${url}, BUT the post data did not match:\n ${advancedTestResults}`);
-    } else if (parameters && this._isInTraffic(url)) {
-      const advancedTestResults = createAdvancedTestResults(url, parameters, this.requests);
-
-      assert.fail(
-        `Traffic named "${name}" found correct URL ${url}, BUT the query parameters did not match:\n`
-        + `${advancedTestResults}`,
-      );
-    } else {
-      assert.fail(
-        `Traffic named "${name}" not found in recorded traffic within ${timeout} seconds.\n`
-        + `Expected url: ${url}.\n`
-        + `Recorded traffic:\n${this._getTrafficDump()}`,
-      );
-    }
+    await seeTraffic.call(this, ...arguments);
   }
 
   /**
@@ -2864,69 +2788,7 @@ class WebDriver extends Helper {
       console.log('* Switch to devtools protocol to use this command by setting devtoolsProtocol: true in the configuration');
       return;
     }
-    if (!this.recordedAtLeastOnce) {
-      throw new Error('Failure in test automation. You use "I.dontSeeTraffic", but "I.startRecordingTraffic" was never called before.');
-    }
-
-    if (!name) {
-      throw new Error('Missing required key "name" in object given to "I.dontSeeTraffic".');
-    }
-
-    if (!url) {
-      throw new Error('Missing required key "url" in object given to "I.dontSeeTraffic".');
-    }
-
-    if (this._isInTraffic(url)) {
-      assert.fail(`Traffic with name "${name}" (URL: "${url}') found, but was not expected to be found.`);
-    }
-  }
-
-  /**
-   * Checks if URL with parameters is part of network traffic. Returns true or false. Internal method for this helper.
-   *
-   * @param url URL to look for.
-   * @param [parameters] Parameters that this URL needs to contain
-   * @return {boolean} Whether or not URL with parameters is part of network traffic.
-   * @private
-   */
-  _isInTraffic(url, parameters) {
-    let isInTraffic = false;
-    this.requests.forEach((request) => {
-      if (isInTraffic) {
-        return; // We already found traffic. Continue with next request
-      }
-
-      if (!request.url.match(new RegExp(url))) {
-        return; // url not found in this request. continue with next request
-      }
-
-      // URL has matched. Now we check the parameters
-
-      if (parameters) {
-        const advancedReport = allParameterValuePairsMatchExtreme(extractQueryObjects(request.url), parameters);
-        if (advancedReport === true) {
-          isInTraffic = true;
-        }
-      } else {
-        isInTraffic = true;
-      }
-    });
-
-    return isInTraffic;
-  }
-
-  /**
-   * Returns all URLs of all network requests recorded so far during execution of test scenario.
-   *
-   * @return {string} List of URLs recorded as a string, separated by new lines after each URL
-   * @private
-   */
-  _getTrafficDump() {
-    let dumpedTraffic = '';
-    this.requests.forEach((request) => {
-      dumpedTraffic += `${request.method} - ${request.url}\n`;
-    });
-    return dumpedTraffic;
+    dontSeeTraffic.call(this, ...arguments);
   }
 }
