+++ lib/helper/networkTraffics/utils.js
@@ -129,9 +129,59 @@ const allRequestPostDataValuePairsMatchExtreme = (RequestPostDataObject, advance
   return success ? true : littleReport;
 };
 
+/**
+ * Returns all URLs of all network requests recorded so far during execution of test scenario.
+ *
+ * @return {string} List of URLs recorded as a string, separated by new lines after each URL
+ * @private
+ */
+function getTrafficDump() {
+  let dumpedTraffic = '';
+  this.requests.forEach((request) => {
+    dumpedTraffic += `${request.method} - ${request.url}\n`;
+  });
+  return dumpedTraffic;
+}
+
+/**
+ * Checks if URL with parameters is part of network traffic. Returns true or false. Internal method for this helper.
+ *
+ * @param url URL to look for.
+ * @param [parameters] Parameters that this URL needs to contain
+ * @return {boolean} Whether or not URL with parameters is part of network traffic.
+ * @private
+ */
+function isInTraffic(url, parameters) {
+  let isInTraffic = false;
+  this.requests.forEach((request) => {
+    if (isInTraffic) {
+      return; // We already found traffic. Continue with next request
+    }
+
+    if (!request.url.match(new RegExp(url))) {
+      return; // url not found in this request. continue with next request
+    }
+
+    // URL has matched. Now we check the parameters
+
+    if (parameters) {
+      const advancedReport = allParameterValuePairsMatchExtreme(extractQueryObjects(request.url), parameters);
+      if (advancedReport === true) {
+        isInTraffic = true;
+      }
+    } else {
+      isInTraffic = true;
+    }
+  });
+
+  return isInTraffic;
+}
+
 module.exports = {
   createAdvancedTestResults,
   extractQueryObjects,
   allParameterValuePairsMatchExtreme,
   allRequestPostDataValuePairsMatchExtreme,
+  getTrafficDump,
+  isInTraffic,
 };
