+++ test/helper/Playwright_test.js
@@ -898,175 +898,6 @@ describe('Playwright', function () {
     });
   });
 
-  describe('#startRecordingTraffic, #seeTraffic, #stopRecordingTraffic, #dontSeeTraffic, #grabRecordedNetworkTraffics', () => {
-    it('should throw error when calling seeTraffic before recording traffics', async () => {
-      try {
-        I.amOnPage('https://codecept.io/');
-        await I.seeTraffic({ name: 'traffics', url: 'https://codecept.io/img/companies/BC_LogoScreen_C.jpg' });
-      } catch (e) {
-        expect(e.message).to.equal('Failure in test automation. You use "I.seeTraffic", but "I.startRecordingTraffic" was never called before.');
-      }
-    });
-
-    it('should throw error when calling seeTraffic but missing name', async () => {
-      try {
-        I.amOnPage('https://codecept.io/');
-        await I.seeTraffic({ url: 'https://codecept.io/img/companies/BC_LogoScreen_C.jpg' });
-      } catch (e) {
-        expect(e.message).to.equal('Missing required key "name" in object given to "I.seeTraffic".');
-      }
-    });
-
-    it('should throw error when calling seeTraffic but missing url', async () => {
-      try {
-        I.amOnPage('https://codecept.io/');
-        await I.seeTraffic({ name: 'https://codecept.io/img/companies/BC_LogoScreen_C.jpg' });
-      } catch (e) {
-        expect(e.message).to.equal('Missing required key "url" in object given to "I.seeTraffic".');
-      }
-    });
-
-    it('should flush the network traffics', async () => {
-      I.startRecordingTraffic();
-      I.amOnPage('https://codecept.io/');
-      I.flushNetworkTraffics();
-      const traffics = await I.grabRecordedNetworkTraffics();
-      expect(traffics.length).to.equal(0);
-    });
-
-    it('should see recording traffics', async () => {
-      I.startRecordingTraffic();
-      I.amOnPage('https://codecept.io/');
-      await I.seeTraffic({ name: 'traffics', url: 'https://codecept.io/img/companies/BC_LogoScreen_C.jpg' });
-    });
-
-    it('should not see recording traffics', async () => {
-      I.startRecordingTraffic();
-      I.amOnPage('https://codecept.io/');
-      I.stopRecordingTraffic();
-      await I.dontSeeTraffic({ name: 'traffics', url: 'https://codecept.io/img/companies/BC_LogoScreen_C.jpg' });
-    });
-
-    it('should not see recording traffics using regex url', async () => {
-      I.startRecordingTraffic();
-      I.amOnPage('https://codecept.io/');
-      I.stopRecordingTraffic();
-      await I.dontSeeTraffic({ name: 'traffics', url: /BC_LogoScreen_C.jpg/ });
-    });
-
-    it('should throw error when calling dontSeeTraffic but missing name', async () => {
-      I.startRecordingTraffic();
-      I.amOnPage('https://codecept.io/');
-      I.stopRecordingTraffic();
-      try {
-        await I.dontSeeTraffic({ url: 'https://codecept.io/img/companies/BC_LogoScreen_C.jpg' });
-      } catch (e) {
-        expect(e.message).to.equal('Missing required key "name" in object given to "I.dontSeeTraffic".');
-      }
-    });
-
-    it('should throw error when calling dontSeeTraffic but missing url', async () => {
-      I.startRecordingTraffic();
-      I.amOnPage('https://codecept.io/');
-      I.stopRecordingTraffic();
-      try {
-        await I.dontSeeTraffic({ name: 'traffics' });
-      } catch (e) {
-        expect(e.message).to.equal('Missing required key "url" in object given to "I.dontSeeTraffic".');
-      }
-    });
-
-    it('should mock traffics', async () => {
-      await I.amOnPage('/form/fetch_call');
-      await I.mockTraffic('https://reqres.in/api/comments/1', '{"name": "this was mocked" }');
-      await I.startRecordingTraffic();
-      await I.click('GET COMMENTS');
-      await I.see('this was mocked');
-
-      await I.mockTraffic('https://reqres.in/api/comments/1', '{"name": "this was another mocked" }');
-      await I.click('GET COMMENTS');
-      await I.see('this was another mocked');
-
-      const traffics = await I.grabRecordedNetworkTraffics();
-      await I.grabRecordedNetworkTraffics();
-      expect(traffics[0].url).to.equal('https://reqres.in/api/comments/1');
-      expect(traffics[0].response.status).to.equal(200);
-      expect(traffics[0].response.body).to.contain({ name: 'this was mocked' });
-
-      expect(traffics[1].url).to.equal('https://reqres.in/api/comments/1');
-      expect(traffics[1].response.status).to.equal(200);
-      expect(traffics[1].response.body).to.contain({ name: 'this was another mocked' });
-    });
-
-    it('should block traffics using a list of urls', async () => {
-      I.blockTraffic(['https://reqres.in/api/*', 'https://reqres.in/api/comments/*']);
-      I.amOnPage('/form/fetch_call');
-      I.startRecordingTraffic();
-      I.click('GET COMMENTS');
-      I.see('Can not load data!');
-    });
-
-    it('should block traffics of a given url', async () => {
-      I.blockTraffic('https://reqres.in/api/comments/*');
-      I.amOnPage('/form/fetch_call');
-      I.startRecordingTraffic();
-      I.click('GET COMMENTS');
-      I.see('Can not load data!');
-    });
-
-    it('should check traffics with more advanced params', async () => {
-      await I.startRecordingTraffic();
-      await I.amOnPage('https://openai.com/blog/chatgpt');
-      const traffics = await I.grabRecordedNetworkTraffics();
-
-      for (const traffic of traffics) {
-        if (traffic.url.includes('&width=')) {
-          // new URL object
-          const currentUrl = new URL(traffic.url);
-
-          // get access to URLSearchParams object
-          const searchParams = currentUrl.searchParams;
-
-          await I.seeTraffic({
-            name: 'sentry event',
-            url: currentUrl.origin + currentUrl.pathname,
-            parameters: searchParams,
-          });
-
-          break;
-        }
-      }
-    });
-
-    it('should check traffics with more advanced post data', async () => {
-      I.amOnPage('https://openai.com/blog/chatgpt');
-      I.startRecordingTraffic();
-      await I.seeTraffic({
-        name: 'event',
-        url: 'https://cloudflareinsights.com/cdn-cgi/rum',
-        requestPostData: {
-          st: 2,
-        },
-      });
-    });
-
-    it('should show error when advanced post data are not matching', async () => {
-      I.amOnPage('https://openai.com/blog/chatgpt');
-      I.startRecordingTraffic();
-      try {
-        await I.seeTraffic({
-          name: 'event',
-          url: 'https://cloudflareinsights.com/cdn-cgi/rum',
-          requestPostData: {
-            st: 3,
-          },
-        });
-      } catch (e) {
-        expect(e.message).to.contain('actual value: "2"');
-      }
-    });
-  });
-
   describe('#startRecordingWebSocketMessages, #grabWebSocketMessages, #stopRecordingWebSocketMessages', () => {
     it('should throw error when calling grabWebSocketMessages before startRecordingWebSocketMessages', () => {
       if (process.env.BROWSER === 'firefox') this.skip();
